// packages/ts/src/types/ast.ts
export interface Block extends ts.Node {
	readonly kind: ts.SyntaxKind.Block;
	readonly parent:
		| AccessorDeclaration
		| ArrowFunction
		| CatchClause
		| ClassStaticBlockDeclaration
		| ConstructorDeclaration
		| FunctionDeclaration
		| FunctionExpression
		| MethodDeclaration
		| StatementParent
		| TryStatement;
	readonly statements: ts.NodeArray<Statement>;
}
// ...
export type CaseOrDefaultClause = CaseClause | DefaultClause;
// ...
export interface Decorator extends ts.Node {
	readonly expression: LeftHandSideExpression;
	readonly kind: ts.SyntaxKind.Decorator;
	readonly parent:
		| AccessorDeclaration
		| ClassDeclaration
		| ClassExpression
		| ConstructorDeclaration
		| EnumDeclaration
		| ExportAssignment
		| ExportDeclaration
		| FunctionDeclaration
		| ImportDeclaration
		| ImportEqualsDeclaration
		| IndexSignatureDeclaration
		| InterfaceDeclaration
		| JSDocNamespaceDeclaration
		| MethodDeclaration
		| ModuleDeclaration
		| NamespaceDeclaration
		| ParameterDeclaration
		| PropertyDeclaration
		| TypeAliasDeclaration
		| VariableStatement;
}
// ...
export interface ExportAssignment extends ts.Node {
	readonly expression: Expression;
	readonly isExportEquals?: boolean;
	readonly kind: ts.SyntaxKind.ExportAssignment;
	readonly modifiers?: ts.NodeArray<ModifierLike>;
	readonly name?: Identifier | NumericLiteral | StringLiteral;
	readonly parent: StatementParent;
}
export interface ExportDeclaration extends ts.Node {
	readonly isTypeOnly: boolean;
	readonly kind: ts.SyntaxKind.ExportDeclaration;
	readonly modifiers?: ts.NodeArray<ModifierLike>;
	readonly parent: StatementParent;
	/** Will not be assigned in the case of `export * from "foo";` */
	readonly exportClause?: NamedExportBindings;
	/** If this is not a StringLiteral it will be a grammar error. */
	readonly attributes?: ImportAttributes;
	readonly moduleSpecifier?: Expression;
	readonly name?: Identifier | NumericLiteral | StringLiteral;
}
export interface ExportKeyword extends ts.Node {
	readonly kind: ts.SyntaxKind.ExportKeyword;
	readonly parent: ModifierParent;
}
export interface ExportSpecifier extends ts.Node {
	readonly isTypeOnly: boolean;
	readonly kind: ts.SyntaxKind.ExportSpecifier;
	readonly name: ModuleExportName;
	readonly parent: NamedExports;
	readonly propertyName?: ModuleExportName;
}
export type Expression =
	| ArrowFunction
	| AsExpression
	| BinaryExpression
	| CommaListExpression
	| ConditionalExpression
	| JsxClosingFragment
	| JsxExpression
	| JsxOpeningElement
	| JsxOpeningFragment
	| OmittedExpression
	| SatisfiesExpression
	| SpreadElement
	| SyntheticExpression
	| UnaryExpression
	| YieldExpression;
// ...
export type FunctionOrConstructorTypeNodeBase =
	| ConstructorTypeNode
	| FunctionTypeNode;
// ...
export interface Identifier extends ts.Node {
	readonly kind: ts.SyntaxKind.Identifier;
	readonly parent:
		| ArrayLiteralExpression
		| ArrowFunction
		| AsExpression
		| AwaitExpression
		| BinaryExpression
		| BindingElement
		| BreakStatement
		| CallExpression
		| CaseClause
		| ClassExpression
		| ClassStaticBlockDeclaration
		| CommaListExpression
		| ComputedPropertyName
		| ConditionalExpression
		| ConstructorDeclaration
		| ContinueStatement
		| DeclarationStatement
		| Decorator
		| DeleteExpression
		| ElementAccessExpression
		| EnumMember
		| ExportSpecifier
		| ExpressionStatement
		| ExternalModuleReference
		| FunctionExpression
		| FunctionOrConstructorTypeNodeBase
		| IfStatement
		| ImportAttribute
		| ImportClause
		| ImportDeclaration
		| ImportSpecifier
		| IterationStatement
		| JSDocFunctionType
		| JSDocLink
		| JSDocLinkCode
		| JSDocLinkPlain
		| JSDocMemberName
		| JSDocNamespaceDeclaration
		| JSDocPropertyLikeTag
		| JSDocReturnTag
		| JSDocTemplateTag
		| JsxAttributeLike
		| JsxClosingElement
		| JsxExpression
		| JsxNamespacedName
		| JsxOpeningElement
		| JsxSelfClosingElement
		| JsxTagNamePropertyAccess
		| LabeledStatement
		| MetaProperty
		| MethodDeclaration
		| NamedTupleMember
		| NamespaceDeclaration
		| NamespaceExport
		| NamespaceImport
		| NewExpression
		| NodeWithTypeArguments
		| NonNullExpression
		| ParameterDeclaration
		| ParenthesizedExpression
		| PartiallyEmittedExpression
		| PostfixUnaryExpression
		| PrefixUnaryExpression
		| PropertyAccessExpression
		| PropertyAssignment
		| PropertyDeclaration
		| QualifiedName
		| ReturnStatement
		| SatisfiesExpression
		| SemicolonClassElement
		| ShorthandPropertyAssignment
		| SpreadAssignment
		| SpreadElement
		| SwitchStatement
		| TaggedTemplateExpression
		| TemplateSpan
		| ThrowStatement
		| TypeAssertion
		| TypeElement
		| TypeOfExpression
		| TypeParameterDeclaration
		| TypePredicateNode
		| VariableDeclaration
		| VoidExpression
		| WithStatement
		| YieldExpression;

	/**
	 * Prefer to use `id.unescapedText`. (Note: This is available only in services, not internally to the TypeScript compiler.)
	 * Text of identifier, but if the identifier begins with two underscores, this will begin with three.
	 */
	readonly escapedText: ts.__String;
	readonly text: string;
}
export interface IfStatement extends ts.Node {
	readonly elseStatement?: Statement;
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.IfStatement;
	readonly parent: StatementParent;
	readonly thenStatement: Statement;
}
export interface ImportAttribute extends ts.Node {
	readonly kind: ts.SyntaxKind.ImportAttribute;
	readonly name: ImportAttributeName;
	readonly parent: ImportAttributes;
	readonly value: Expression;
}
export type ImportAttributeName = Identifier | StringLiteral;
export interface ImportAttributes extends ts.Node {
	readonly elements: ts.NodeArray<ImportAttribute>;
	readonly kind: ts.SyntaxKind.ImportAttributes;
	readonly multiLine?: boolean;
	readonly parent: ExportDeclaration | ImportDeclaration | ImportTypeNode;
	readonly token: ts.SyntaxKind.AssertKeyword | ts.SyntaxKind.WithKeyword;
}
export interface ImportClause extends ts.Node {
	readonly kind: ts.SyntaxKind.ImportClause;
	readonly parent: ImportDeclaration;
	/** @deprecated Use `phaseModifier` instead */
	readonly isTypeOnly: boolean;
	readonly name?: Identifier;
	readonly namedBindings?: NamedImportBindings;
	readonly phaseModifier: ImportPhaseModifierSyntaxKind | undefined;
}
export interface ImportDeclaration extends ts.Node {
	readonly importClause?: ImportClause;
	readonly kind: ts.SyntaxKind.ImportDeclaration;
	readonly modifiers?: ts.NodeArray<ModifierLike>;
	readonly parent: StatementParent;
	/** If this is not a StringLiteral it will be a grammar error. */
	readonly attributes?: ImportAttributes;
	readonly moduleSpecifier: Expression;
}
// ...
export type ImportPhaseModifierSyntaxKind =
	| ts.SyntaxKind.DeferKeyword
	| ts.SyntaxKind.TypeKeyword;
// ...
export type IterationStatement =
	| DoStatement
	| ForInStatement
	| ForOfStatement
	| ForStatement
	| WhileStatement;
// ...
export interface LabeledStatement extends ts.Node {
	readonly kind: ts.SyntaxKind.LabeledStatement;
	readonly label: Identifier;
	readonly parent: StatementParent;
	readonly statement: Statement;
}
// ...
export type Modifier =
	| AbstractKeyword
	| AccessorKeyword
	| AsyncKeyword
	| ConstKeyword
	| DeclareKeyword
	| DefaultKeyword
	| ExportKeyword
	| InKeyword
	| OutKeyword
	| OverrideKeyword
	| PrivateKeyword
	| ProtectedKeyword
	| PublicKeyword
	| ReadonlyKeyword
	| StaticKeyword;
export type ModifierLike = Decorator | Modifier;
export interface ModuleBlock extends ts.Node {
	readonly kind: ts.SyntaxKind.ModuleBlock;
	readonly parent:
		| Block
		| CaseOrDefaultClause
		| IfStatement
		| IterationStatement
		| LabeledStatement
		| ModuleDeclaration
		| NamespaceBody
		| SourceFile
		| WithStatement;
	readonly statements: ts.NodeArray<Statement>;
}
// ...
export type NamedExportBindings = NamedExports | NamespaceExport;
export interface NamedExports extends ts.Node {
	readonly elements: ts.NodeArray<ExportSpecifier>;
	readonly kind: ts.SyntaxKind.NamedExports;
	readonly parent: ExportDeclaration;
}
export type NamedImportBindings = NamedImports | NamespaceImport;
// ...
export interface NumericLiteral extends ts.Node {
	hasExtendedUnicodeEscape?: boolean;
	isUnterminated?: boolean;
	readonly kind: ts.SyntaxKind.NumericLiteral;
	readonly parent:
		| ClassStaticBlockDeclaration
		| ConstructorDeclaration
		| FunctionOrConstructorTypeNodeBase
		| JSDocFunctionType
		| LeftHandSideExpressionParent
		| LiteralTypeNode
		| MethodDeclaration
		| SemicolonClassElement
		| TypeElement;
	text: string;
}
// ...
export interface SourceFile extends ts.Node {
	readonly kind: ts.SyntaxKind.SourceFile;
	// parent is actually undefined, see comment for NullNode
	amdDependencies: readonly ts.AmdDependency[];
	readonly endOfFileToken: Token<ts.SyntaxKind.EndOfFileToken, SourceFile>;
	fileName: string;
	isDeclarationFile: boolean;
	languageVariant: ts.LanguageVariant;
	libReferenceDirectives: readonly ts.FileReference[];
	moduleName?: string;
	readonly parent: NullNode;
	referencedFiles: readonly ts.FileReference[];
	readonly statements: ts.NodeArray<Statement>;
	text: string;
	typeReferenceDirectives: readonly ts.FileReference[];

	/**
	 * lib.d.ts should have a reference comment like
	 *
	 *  /// &lt;reference no-default-lib="true"/>
	 *
	 * If any other file has this comment, it signals not to include lib.d.ts
	 * because this containing file is intended to act as a default library.
	 */
	hasNoDefaultLib: boolean;
	languageVersion: ts.ScriptTarget;

	/**
	 * When `module` is `Node16` or `NodeNext`, this field controls whether the
	 * source file in question is an ESNext-output-format file, or a CommonJS-output-format
	 * module. This is derived by the module resolver as it looks up the file, since
	 * it is derived from either the file extension of the module, or the containing
	 * `package.json` context, and affects both checking and emit.
	 *
	 * It is _public_ so that (pre)transformers can set this field,
	 * since it switches the builtin `node` module transform. Generally speaking, if unset,
	 * the field is treated as though it is `ModuleKind.CommonJS`.
	 *
	 * Note that this field is only set by the module resolution process when
	 * `moduleResolution` is `Node16` or `NodeNext`, which is implied by the `module` setting
	 * of `Node16` or `NodeNext`, respectively, but may be overriden (eg, by a `moduleResolution`
	 * of `node`). If so, this field will be unset and source files will be considered to be
	 * CommonJS-output-format by the node module transformer and type checker, regardless of extension or context.
	 */
	getLineAndCharacterOfPosition(pos: number): ts.LineAndCharacter;
	getLineEndOfPosition(pos: number): number;
	getLineStarts(): readonly number[];
	getPositionOfLineAndCharacter(line: number, character: number): number;
	impliedNodeFormat?: ts.ResolutionMode;
	update(newText: string, textChangeRange: ts.TextChangeRange): SourceFile;
}
// ...
export interface StringLiteral extends ts.Node {
	hasExtendedUnicodeEscape?: boolean;
	isUnterminated?: boolean;
	readonly kind: ts.SyntaxKind.StringLiteral;
	readonly parent:
		| ArrayLiteralExpression
		| ArrowFunction
		| AsExpression
		| AwaitExpression
		| BinaryExpression
		| BindingElement
		| CallExpression
		| CaseClause
		| ClassStaticBlockDeclaration
		| CommaListExpression
		| ComputedPropertyName
		| ConditionalExpression
		| ConstructorDeclaration
		| Decorator
		| DeleteExpression
		| ElementAccessExpression
		| EnumMember
		| ExportAssignment
		| ExportDeclaration
		| ExportSpecifier
		| ExpressionStatement
		| ExpressionWithTypeArguments
		| ExternalModuleReference
		| FunctionOrConstructorTypeNodeBase
		| IfStatement
		| ImportAttribute
		| ImportDeclaration
		| ImportSpecifier
		| IterationStatement
		| JSDocFunctionType
		| JsxAttributeLike
		| JsxExpression
		| LiteralTypeNode
		| MethodDeclaration
		| ModuleDeclaration
		| NamespaceExport
		| NewExpression
		| NonNullExpression
		| ParameterDeclaration
		| ParenthesizedExpression
		| PartiallyEmittedExpression
		| PostfixUnaryExpression
		| PrefixUnaryExpression
		| PropertyAccessExpression
		| PropertyAssignment
		| PropertyDeclaration
		| ReturnStatement
		| SatisfiesExpression
		| SemicolonClassElement
		| ShorthandPropertyAssignment
		| SpreadAssignment
		| SpreadElement
		| SwitchStatement
		| TaggedTemplateExpression
		| TemplateSpan
		| ThrowStatement
		| TypeAssertion
		| TypeElement
		| TypeOfExpression
		| TypeParameterDeclaration
		| VariableDeclaration
		| VoidExpression
		| WithStatement
		| YieldExpression;
	text: string;
}
// ...
export interface WithStatement extends ts.Node {
	readonly expression: Expression;
	readonly kind: ts.SyntaxKind.WithStatement;
	readonly parent: StatementParent;
	readonly statement: Statement;
}
// ...
export type StatementParent =
	| Block
	| CaseOrDefaultClause
	| IfStatement
	| IterationStatement
	| LabeledStatement
	| ModuleBlock
	| SourceFile
	| WithStatement;